% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_tsne.R
\name{predict.tsne}
\alias{predict.tsne}
\title{predict T-SNE}
\usage{
\method{predict}{tsne}(object, data, k = 5, ...)
}
\arguments{
\item{object}{a list from \link{train_tsne}}

\item{data}{a data want to pred}

\item{k}{how many neighbor to use}

\item{...}{Other}
}
\value{
a dataframe for 2 dim
}
\description{
use MDS to setup Y_init , then we can predict the T-SNE
}
