% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tsne.R
\name{predict_tsne}
\alias{predict_tsne}
\title{predict T-SNE}
\usage{
predict_tsne(TSNE, data, k = 5)
}
\arguments{
\item{TSNE}{a list from \link{train_tsne}}

\item{data}{a data want to pred}

\item{k}{how many neighbor to use}
}
\value{
a dataframe for 2 dim
}
\description{
use MDS to setup Y_init , then we can predict the T-SNE
}
\examples{
library(RTSNEpred)

library(ggplot2)
library(magrittr)
set.seed(123)
iris_index <- sample(1:nrow(iris),size = floor(0.8*nrow(iris)))
iris_train <- iris[iris_index,]
iris_test <- iris[-iris_index,]
set.seed(123)
iris_tsne <- train_tsne(iris_train[,1:4])
iris_pred <- predict_tsne(TSNE = iris_tsne,data = iris_test[,1:4],k = 3) \%>\%
  as.data.frame() \%>\%
  cbind(iris_test$Species) \%>\%
  `names<-`(c("x","y","Species"))
iris_tsne \%$\%
  Y \%>\%
  as.data.frame() \%>\%
  cbind(iris_train$Species) \%>\%
  `names<-`(c("x","y","Species")) \%>\%
  ggplot(aes(x=x,y=y,color = Species)) +
  geom_point()

ggplot(iris_pred,aes(x=x,y=y,color = Species)) +
  geom_point()

iris_tsne \%$\%
  Y \%>\%
  as.data.frame() \%>\%
  cbind(iris_train$Species) \%>\%
  `names<-`(c("x","y","Species")) \%>\%
  rbind(iris_pred) \%>\%
  ggplot(aes(x=x,y=y,color = Species)) +
  geom_point()
}
